FileContextMenu: run paste action on menu select
The paste function was moved to the appState object because most file/batch
related functions are stored in this object. Maybe we should move this
elsewhere ? There is also a Batch state (which is defined in its own transfers dir ?!)...

We should do some cleanup here!
